reactLiftingState
